requirements.txt
# Core dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# Document processing
unstructured[all-docs]==0.11.6
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.1.0

# Vector database and embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Google Gemini
google-generativeai==0.3.2

# Utilities
pandas==2.1.4
numpy==1.26.2
